Change timeout to be a long as per all other timeouts (FailOnTimeout, Test(timeout=).

+++ src/main/java/org/junit/rules/Timeout.java
@@ -34,12 +34,12 @@ import org.junit.runners.model.Statement;
  * </pre>
  */
 public class Timeout implements TestRule {
-	private final int fMillis;
+	private final long fMillis;
 
 	/**
 	 * @param millis the millisecond timeout
 	 */
-	public Timeout(int millis) {
+	public Timeout(long millis) {
 		fMillis= millis;
 	}
